Pack PinState into a u16 and edit it in place
Pins only ever carry up to 8 wires (PinBitCount tops out at Bit8), so the
state word halves: bit values in the low byte, tri-state flags in the high
byte. set_bit/toggle_bit now clear and set exactly their wire's bits in
the packed word instead of rebuilding both halves through fresh locals,
and every hot-path update already mutates arena entries directly -- PinState
is a 2-byte Copy with no heap anywhere near the stepping loop. RAM and dot
display outputs drive from_parts(low byte, 0) since cells hold random u32s
whose upper lanes no reader could reach.

+++ tests/audio_test.rs
@@ -152,8 +152,8 @@ mod end_to_end {
 	/// (id 0, 4-bit); the sim arm reads them in that order.
 	fn drive(sim: &mut Simulator, audio: &mut SimAudio, pitch: u32, volume: u32) {
 		let inputs = [
-			ExternalInput { address: PinAddress::new(1, 1), state: PinState::from_raw(pitch) },
-			ExternalInput { address: PinAddress::new(1, 0), state: PinState::from_raw(volume) },
+			ExternalInput { address: PinAddress::new(1, 1), state: PinState::from_raw(pitch as u16) },
+			ExternalInput { address: PinAddress::new(1, 0), state: PinState::from_raw(volume as u16) },
 		];
 		sim.run_simulation_step(&inputs, audio);
 	}
